requirements.txt
# requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.4
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.6
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-genai==1.62.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.3.1
marshmallow==3.26.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.4.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.21.0
openai==1.66.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMatting==1.1.13
PyMuPDF==1.25.5
pyparsing==3.2.3
pyreadline3==3.5.4
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.19
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tifffile==2025.3.30
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.32.1
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
